Basic Calculator: A Personal Python Project.md
In this repository, I proudly present my personal project: a Basic Calculator crafted in Python. This calculator embodies a functional solution to perform essential arithmetic operations such as addition, subtraction, multiplication, and division. Designed meticulously for command-line interaction, it offers a streamlined, user-friendly interface that allows for intuitive computation.

This project not only showcases my skills in Python programming but also reflects my commitment to creating practical, user-centric applications. Through the development of this calculator, I aimed to fuse simplicity with functionality, providing a tool that is both accessible and useful in everyday tasks.